getWindowSize hook
